fix(tests): Update tests to use MySQL as default SqlMode
After commit f2e4b9a9 changed the default SqlMode from SQLite to MySQL
for SQLLogicTest compatibility, several tests were left assuming SQLite
was still the default. This fix updates all affected tests.

Changes:
- tests/regression/issue_2783.rs: Explicitly set SQLite mode for tests
  that verify SQLite-specific integer division and CAST truncation
- tests/regression/issue_2852.rs: Explicitly set SQLite mode for tests
  that verify SQLite integer division behavior
- crates/vibesql-executor/src/evaluator/operators/arithmetic/mod.rs:
  Unit tests for division that test SQLite-specific behavior
- crates/vibesql-executor/tests/conversion_function_tests.rs:
  Update CAST test to expect MySQL rounding behavior (43 not 42)
- crates/vibesql-types/src/sql_mode/mod.rs:
  Rename test_default_is_sqlite to test_default_is_mysql
- crates/vibesql-types/src/sql_mode/types.rs:
  Update test_has_decimal_type and test_uses_dynamic_typing

Closes #2872

+++ crates/vibesql-executor/tests/conversion_function_tests.rs
@@ -397,10 +397,10 @@ fn test_cast_integer_to_double() {
 
 #[test]
 fn test_cast_double_to_integer() {
-    // Default mode is SQLite which truncates toward zero when casting to INTEGER
-    // MySQL mode would round 42.9 to 43, but SQLite truncates to 42
+    // Default mode is MySQL which rounds when casting to INTEGER
+    // MySQL mode rounds 42.9 to 43 (SQLite would truncate to 42)
     let result = eval_function("CAST", vec![double_lit(42.9), varchar_lit("INTEGER")]);
-    assert_eq!(result, vibesql_types::SqlValue::Integer(42));
+    assert_eq!(result, vibesql_types::SqlValue::Integer(43));
 }
 
 #[test]
